Move vg call to its own subcommand cpp
